+++ asapdiscovery-data/asapdiscovery/data/tests/test_sets.py
@@ -4,6 +4,7 @@ from asapdiscovery.data.schema_v2.sets import CompoundMultiStructure
 
 
 def test_multi_structure_from_pairs(ligands, complexes):
+    # I could use the pairwise selector but that makes it an integration test, right?
     pairs = [
         CompoundStructurePair(ligand=ligand, complex=complex)
         for ligand, complex in product(ligands, complexes)
@@ -11,4 +12,4 @@ def test_multi_structure_from_pairs(ligands, complexes):
     assert len(pairs) == 40
     multi_structures = CompoundMultiStructure.from_pairs(pairs)
 
-    assert len(multi_structures) == 10
+    assert len(multi_structures) == 4
